refactoring bit interleave out of Morton so it can be used with Hilbert
